feat: :sparkles: Each service request is assigned to a single H3 resolution level 8 hexagon.
New function added to the main class to assign service requests to hexagons. A test to assert that the resultant df is equal to sample `sr_hex.csv.gz` is also added to test_main.py.

Issue: The test fails since there is a 0.00308% difference between the two dataframes. This happens when boundary coordinates are assigned to adjacent hexagons.